detail list + context menu

+++ chatclientadmin/src/main/java/com/hcmus/ui/table/ContextMenu.java
@@ -1,22 +1,58 @@
 package com.hcmus.ui.table;
 
 import javax.swing.*;
+import java.awt.event.ActionListener;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
-import java.sql.Connection;
+import java.util.List;
 
 public class ContextMenu extends JPopupMenu {
     private JMenuItem editItem;
     private JMenuItem deleteItem;
+    private JMenuItem blockItem;
+    private JMenu detailItem;
+    private JMenuItem loginHisItem;
+    private JMenuItem friendListItem;
+    private JMenuItem memListItem;
+    private JMenuItem adminListItem;
     private JTable table;
+    private List<String> options;
 
-    private Connection conn;
-    public ContextMenu(JTable table) {
+    public ContextMenu(JTable table, List<String> options) {
         this.table = table;
-        editItem = new JMenuItem("Edit");
-        deleteItem = new JMenuItem("Delete");
-        add(editItem);
-        add(deleteItem);
+        this.options = options;
+        if (options.contains("Edit")) {
+            editItem = new JMenuItem("Edit");
+            add(editItem);
+        }
+        if (options.contains("Delete")) {
+            deleteItem = new JMenuItem("Delete");
+            add(deleteItem);
+        }
+        if (options.contains("Block")) {
+            blockItem = new JMenuItem("Block");
+            add(blockItem);
+        }
+        if (options.contains("Detail")) {
+            detailItem = new JMenu("Detail");
+            if (options.contains("Login History")) {
+                loginHisItem = new JMenuItem("Login History");
+                detailItem.add(loginHisItem);
+            }
+            if (options.contains("Friend List")) {
+                friendListItem = new JMenuItem("Friend List");
+                detailItem.add(friendListItem);
+            }
+            if (options.contains("Member List")) {
+                memListItem = new JMenuItem("Member List");
+                detailItem.add(memListItem);
+            }
+            if (options.contains("Admin List")) {
+                adminListItem = new JMenuItem("Admin List");
+                detailItem.add(adminListItem);
+            }
+            add(detailItem);
+        }
         table.addMouseListener(new MouseAdapter() {
             @Override
             public void mousePressed(MouseEvent e) {
@@ -50,12 +86,50 @@ public class ContextMenu extends JPopupMenu {
             table.setRowSelectionInterval(row, row);
             editItem.setEnabled(true);
             deleteItem.setEnabled(true);
+            blockItem.setEnabled(true);
+            detailItem.setEnabled(true);
         } else {
             table.clearSelection();
             editItem.setEnabled(false);
             deleteItem.setEnabled(false);
+            blockItem.setEnabled(false);
+            detailItem.setEnabled(false);
         }
         show(table, e.getX(), e.getY());
     }
+
+    public int getSelectedRowIndex() {
+        return table.getSelectedRow();
+    }
+
+    // Add action listener for each menu item
+    public void addEditListener(ActionListener listener) {
+        editItem.addActionListener(listener);
+    }
+
+    public void addDeleteListener(ActionListener listener) {
+        deleteItem.addActionListener(listener);
+    }
+
+    public void addBlockListener(ActionListener listener) {
+        blockItem.addActionListener(listener);
+    }
+
+    public void addLoginHisListener(ActionListener listener) {
+        loginHisItem.addActionListener(listener);
+    }
+
+    public void addFriendListListener(ActionListener listener) {
+        friendListItem.addActionListener(listener);
+    }
+
+    public void addMemListListener(ActionListener listener) {
+        memListItem.addActionListener(listener);
+    }
+
+    public void addAdminListListener(ActionListener listener) {
+        adminListItem.addActionListener(listener);
+    }
+
 }
